assert_is_empty_model {assertive.models}	R Documentation

Is the input the empty model?

Description

Checks to see if the input is the empty model.

Usage

assert_is_empty_model(x, severity = getOption("assertive.severity",
  "stop"))

assert_is_non_empty_model(x, severity = getOption("assertive.severity",
  "stop"))

is_empty_model(x, .xname = get_name_in_parent(x))

is_non_empty_model(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_[non_]empty_model returns TRUE if the input is an [non] empty model. (has_terms is used to determine that a variable is a model object.) The model is considered empty if there are no factors and no intercept. The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

See Also

is.empty.model and is_empty.

Examples

# empty models have no intercept and no factors
an_empty_model <- lm(uptake ~ 0, CO2)
is_empty_model(an_empty_model)

a_model_with_an_intercept <- lm(uptake ~ 1, CO2)
a_model_with_factors <- lm(uptake ~ conc * Type, CO2)
is_non_empty_model(a_model_with_an_intercept)
is_non_empty_model(a_model_with_factors)

assertive.base::dont_stop(assert_is_empty_model(a_model_with_factors))

[Package assertive.models version 0.0-2 Index]